Increase DiceWin performance
This commit modify the dicewin script to use dataframes and arrays to
store data and reduces the time and memory used to read a file. Also, it
adds support to a generic data file, with the first line holding data
labels. Furthermore, it adds the possibility to customize the plots
through the new icon on the toolbar